pyproject.toml
[project]
name = "uitwiki-chatbot"
version = "0.1.0"
description = "uitWiki Chatbot Service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.8.0.post1",
    "fastapi[standard]>=0.115.0",
    "google-generativeai==0.7.2",
    "langchain-community>=0.3.1",
    "langchain-google-genai>=2.0.0",
    "langchain-huggingface>=0.1.0",
    "langchain>=0.3.1",
    "pydantic-settings>=2.5.2",
    "pypdf2>=3.0.1",
]
